Run garden as systemd service when systemd running
We need to do that in order to tell systemd to not
touch the cgroups we create and manage by setting
the Delegate=yes property for the garden service.
Otherwise systemd will move container processes out
of their respective cgroups since it does not expect
that we are managing our own cgroups.

[#162329293]

Co-authored-by: Julia Nedialkova <julianedialkova@hotmail.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>